refactor(scanner): decompose scanner.py god module — PR 1/2 (#199)
- Extract session_metrics.py (~135 lines): calculate_day_metrics_from_aggs
  and calculate_day_metrics as module-level functions
- Extract scan_enrichment.py (~185 lines): _SECTOR_ETF_MAP/_SECTOR_ETF_SYMBOLS
  and both batch enrichment functions
- Move run_pre_market_scan body into pre_market_scan.py; _run adapter now
  calls the local function directly
- Move run_oversold_bounce_scan body into oversold_bounce_scan.py; same pattern
- Reduce scanner.py from 1,033 → 206 lines as a thin delegation facade
- Lazy function-body imports in scan runners avoid the load-time cycle
- load_ranker_config/get_volume_forecast kept as noqa:F401 patch targets

Zero import sites changed. All 5 callers reference only ScannerService.
All existing tests pass unchanged.
